pyproject.toml
[project]
name = "dev-snowflake-dbt-analytics-platform"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.7.0,<2.0.0",
    "dbt-duckdb>=1.7.0,<2.0.0",
    "dbt-snowflake>=1.7.0,<2.0.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "snowflake-connector-python>=4.2.0",
]
